+++ Tutorials/03_handwriting_recognition/train.py
@@ -16,7 +16,7 @@ from mltu.tensorflow.metrics import CWERMetric
 from model import train_model
 from configs import ModelConfigs
 
-import stow
+import os
 import tarfile
 from tqdm import tqdm
 from urllib.request import urlopen
@@ -34,17 +34,17 @@ def download_and_unzip(url, extract_to='Datasets', chunk_size=1024*1024):
     zipfile = ZipFile(BytesIO(data))
     zipfile.extractall(path=extract_to)
 
-dataset_path = stow.join('Datasets', 'IAM_Words')
-if not stow.exists(dataset_path):
+dataset_path = os.path.join('Datasets', 'IAM_Words')
+if not os.path.exists(dataset_path):
     download_and_unzip('https://git.io/J0fjL', extract_to='Datasets')
 
-    file = tarfile.open(stow.join(dataset_path, "words.tgz"))
-    file.extractall(stow.join(dataset_path, "words"))
+    file = tarfile.open(os.path.join(dataset_path, "words.tgz"))
+    file.extractall(os.path.join(dataset_path, "words"))
 
 dataset, vocab, max_len = [], set(), 0
 
 # Preprocess the dataset by the specific IAM_Words dataset file structure
-words = open(stow.join(dataset_path, "words.txt"), "r").readlines()
+words = open(os.path.join(dataset_path, "words.txt"), "r").readlines()
 for line in tqdm(words):
     if line.startswith("#"):
         continue
@@ -58,8 +58,8 @@ for line in tqdm(words):
     file_name = line_split[0] + ".png"
     label = line_split[-1].rstrip('\n')
 
-    rel_path = stow.join(dataset_path, "words", folder1, folder2, file_name)
-    if not stow.exists(rel_path):
+    rel_path = os.path.join(dataset_path, "words", folder1, folder2, file_name)
+    if not os.path.exists(rel_path):
         continue
 
     dataset.append([rel_path, label])
@@ -130,5 +130,5 @@ model.fit(
 )
 
 # Save training and validation datasets as csv files
-train_data_provider.to_csv(stow.join(configs.model_path, 'train.csv'))
-val_data_provider.to_csv(stow.join(configs.model_path, 'val.csv'))
+train_data_provider.to_csv(os.path.join(configs.model_path, 'train.csv'))
+val_data_provider.to_csv(os.path.join(configs.model_path, 'val.csv'))