+++ mltu/dataProvider.py
@@ -14,19 +14,6 @@ logger = logging.getLogger(__name__)
 logger.setLevel(logging.INFO)
 
 class DataProvider:
-    """ Standardised object for providing data to a TensorFlow model while training.
-
-    Args:
-        dataset (str, list, pd.DataFrame): Path to dataset, list of data or pandas dataframe of data.
-        data_preprocessors (list): List of data preprocessors. (e.g. [read image, read audio, etc.])
-        batch_size (int, optional): The number of samples to include in each batch. Defaults to 4.
-        shuffle (bool): Whether to shuffle the data. Defaults to True.
-        initial_epoch (int): The initial epoch. Defaults to 1.
-        augmentors (list, optional): List of augmentor functions. Defaults to None.
-        transformers (list, optional): List of transformer functions. Defaults to None.
-        skip_validation (bool, optional): Whether to skip validation. Defaults to False.
-        limit (int, optional): Limit the number of samples in the dataset. Defaults to None.
-    """
     def __init__(
         self, 
         dataset: typing.Union[str, list, pd.DataFrame],
@@ -39,7 +26,19 @@ class DataProvider:
         skip_validation: bool = True,
         limit: int = None,
         ) -> None:
-        super().__init__()
+        """ Standardised object for providing data to a model while training.
+
+        Attributes:
+            dataset (str, list, pd.DataFrame): Path to dataset, list of data or pandas dataframe of data.
+            data_preprocessors (list): List of data preprocessors. (e.g. [read image, read audio, etc.])
+            batch_size (int): The number of samples to include in each batch. Defaults to 4.
+            shuffle (bool): Whether to shuffle the data. Defaults to True.
+            initial_epoch (int): The initial epoch. Defaults to 1.
+            augmentors (list, optional): List of augmentor functions. Defaults to None.
+            transformers (list, optional): List of transformer functions. Defaults to None.
+            skip_validation (bool, optional): Whether to skip validation. Defaults to True.
+            limit (int, optional): Limit the number of samples in the dataset. Defaults to None.
+        """
         self._dataset = self.validate(dataset, skip_validation, limit)
         self._data_preprocessors = data_preprocessors
         self._batch_size = batch_size